Namespace part of UserPolicy
