+++ app/policies/user_policy.rb
@@ -6,16 +6,4 @@ class UserPolicy < ApplicationPolicy
 
     user == record
   end
-
-  def create?
-    raise Pundit::NotAuthorizedError, query: :authenticate? unless user
-
-    record.brand.nil?
-  end
-
-  def destroy?
-    raise Pundit::NotAuthorizedError, query: :authenticate? unless user
-
-    user.brand == record.brand
-  end
 end
